Se agrega cuandro de dialogo en el boton de Eliminar Usuario de las opciones del Administrador, parq que confirme la eliminacion. Se modifico en el mismo metodo el tipo de dato de la variable de id_usuario de int a short

+++ src/main/java/com/edu/konradlorenz/controller/Controlador.java
@@ -47,7 +47,7 @@ public class Controlador {
         return controlPersistencia.traerUsuario(id_user);
     }
 
-    public void eliminarUsuario(int id_usuario) {
+    public void eliminarUsuario(short id_usuario) {
         controlPersistencia.eliminarUsuario(id_usuario);
     }
